test: logging config sources when running jest
