+++ package.json
@@ -35,6 +35,7 @@
     "start:styleguide": "docker-compose run --no-deps -p 6060:6060 app yarn run styleguide",
     "stop:services": "docker-compose down -v",
     "styleguide": "styleguidist server",
+    "pretest": "NODE_ENV=test log-config-sources",
     "test": "jest --runInBand",
     "test:browser": "NODE_CONFIG_ENV=test NODE_ENV=production testcafe \"${BROWSER:-chrome}\" 'test/**/*.browser.js'",
     "test:dependencies": "check-yarn-dupes react react-dom styled-components pubsweet-server",
