+++ server/util/package.json
@@ -1,14 +1,12 @@
 {
   "name": "@elifesciences/util",
   "version": "1.0.0",
-  "description": "",
   "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
   "license": "ISC",
+  "bin": {
+    "log-config-sources": "./logConfigSources.js"
+  },
   "dependencies": {
-    "config": "^3.0.0"
+    "config": "^2.0.1"
   }
 }
